refactor: renaming to `settings.py`
